requirements.txt
aiohttp==1.0.5
async-timeout==3.0.1
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.10.15
chardet==3.0.4
discord==0.0.2
discord.py==0.16.12
idna==2.7
lxml==4.2.5
multidict==4.5.0
numpy==1.15.4
pandas==0.23.4
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.20.1
six==1.11.0
urllib3==1.24.1
websockets==3.4
